SDK-306 move input validation to separate folder
