+++ src/api-client.ts
@@ -29,15 +29,15 @@ import { BatchWriteResponseIO, BatchWriteResponse } from './validation/api/batch
 import { DeleteResponseIO, DeleteResponse } from './validation/api/delete-response';
 import { AddAttachmentResponseIO, AddAttachmentResponse } from './validation/api/add-attachment-response';
 import { UpsertAttachmentResponse, UpsertAttachmentResponseIO } from './validation/api/upsert-attachment-response';
-import { FindFilter } from './validation/find-filter';
+import { FindFilter } from './validation/user-input/find-filter';
 import { ApiFindOptions } from './validation/api/api-find-options';
 import { ApiRecordData } from './validation/api/api-record-data';
-import { RequestOptions } from './validation/request-options';
-import { AttachmentWritableMeta } from './validation/attachment-writable-meta';
+import { RequestOptions } from './validation/user-input/request-options';
+import { AttachmentWritableMeta } from './validation/user-input/attachment-writable-meta';
 import { UpdateAttachmentMetaResponse, UpdateAttachmentMetaResponseIO } from './validation/api/update-attachment-meta-response';
 import { GetAttachmentMetaResponse, GetAttachmentMetaResponseIO } from './validation/api/get-attachment-meta-response';
 import { getFileNameFromHeaders } from './utils';
-import { AttachmentData } from './validation/api/attachment-data';
+import { AttachmentData } from './validation/user-input/attachment-data';
 
 const pjson = require('../package.json');
