+++ src/validation/user-input/attachment-data.ts
