+++ src/validation/user-input/attachment-writable-meta.ts
