+++ tests/unit/storage/add-attachment-tests.ts
@@ -13,7 +13,7 @@ import {
   EMPTY_API_RESPONSE_ATTACHMENT_META,
 } from './common';
 import { InputValidationError, StorageNetworkError } from '../../../src/errors';
-import { COUNTRY_CODE_ERROR_MESSAGE } from '../../../src/validation/country-code';
+import { COUNTRY_CODE_ERROR_MESSAGE } from '../../../src/validation/user-input/country-code';
 import { nockPopApi, getNockedRequestBodyRaw } from '../../test-helpers/popapi-nock';
 import { Storage } from '../../../src/storage';
 import { errorMessageRegExp } from '../../test-helpers/utils';
