+++ tests/unit/storage/update-attachment-meta-tests.ts
@@ -10,10 +10,10 @@ import {
   EMPTY_API_RESPONSE_ATTACHMENT_META,
 } from './common';
 import { InputValidationError, StorageNetworkError } from '../../../src/errors';
-import { COUNTRY_CODE_ERROR_MESSAGE } from '../../../src/validation/country-code';
+import { COUNTRY_CODE_ERROR_MESSAGE } from '../../../src/validation/user-input/country-code';
 import { nockPopApi } from '../../test-helpers/popapi-nock';
 import { Storage } from '../../../src/storage';
-import { AttachmentWritableMeta } from '../../../src/validation/attachment-writable-meta';
+import { AttachmentWritableMeta } from '../../../src/validation/user-input/attachment-writable-meta';
 import { errorMessageRegExp } from '../../test-helpers/utils';
 
 chai.use(chaiAsPromised);
