+++ tests/unit/storage/write-tests.ts
@@ -21,7 +21,7 @@ import {
 import { VALID_REQUEST_OPTIONS, INVALID_REQUEST_OPTIONS } from '../validation/request-options';
 import { Storage, WriteResult } from '../../../src/storage';
 import { InputValidationError, StorageError } from '../../../src/errors';
-import { COUNTRY_CODE_ERROR_MESSAGE } from '../../../src/validation/country-code';
+import { COUNTRY_CODE_ERROR_MESSAGE } from '../../../src/validation/user-input/country-code';
 import { ApiRecordData } from '../../../src/validation/api/api-record-data';
 import { errorMessageRegExp } from '../../test-helpers/utils';
